break[array]: lazy list_contains
