+++ encodings/sequence/src/rules.rs
@@ -3,11 +3,13 @@
 
 use vortex_array::arrays::SliceReduceAdaptor;
 use vortex_array::compute::CastReduceAdaptor;
+use vortex_array::compute::ListContainsElementReduceAdaptor;
 use vortex_array::optimizer::rules::ParentRuleSet;
 
 use crate::SequenceVTable;
 
 pub(crate) static RULES: ParentRuleSet<SequenceVTable> = ParentRuleSet::new(&[
     ParentRuleSet::lift(&CastReduceAdaptor(SequenceVTable)),
+    ParentRuleSet::lift(&ListContainsElementReduceAdaptor(SequenceVTable)),
     ParentRuleSet::lift(&SliceReduceAdaptor(SequenceVTable)),
 ]);
