--- vortex-array/src/expr/exprs/list_contains.rs
@@ -1,429 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-use std::fmt::Formatter;
-use std::ops::BitOr;
-
-use vortex_dtype::DType;
-use vortex_error::VortexResult;
-use vortex_error::vortex_bail;
-use vortex_error::vortex_err;
-use vortex_session::VortexSession;
-
-use crate::ArrayRef;
-use crate::IntoArray;
-use crate::arrays::ConstantArray;
-use crate::compute::list_contains as compute_list_contains;
-use crate::expr::Arity;
-use crate::expr::ChildName;
-use crate::expr::EmptyOptions;
-use crate::expr::ExecutionArgs;
-use crate::expr::ExprId;
-use crate::expr::Expression;
-use crate::expr::StatsCatalog;
-use crate::expr::VTable;
-use crate::expr::VTableExt;
-use crate::expr::and_collect;
-use crate::expr::exprs::binary::gt;
-use crate::expr::exprs::binary::lt;
-use crate::expr::exprs::binary::or;
-use crate::expr::exprs::literal::Literal;
-use crate::expr::exprs::literal::lit;
-use crate::scalar::Scalar;
-
-pub struct ListContains;
-
-impl VTable for ListContains {
-    type Options = EmptyOptions;
-
-    fn id(&self) -> ExprId {
-        ExprId::from("vortex.list.contains")
-    }
-
-    fn serialize(&self, _instance: &Self::Options) -> VortexResult<Option<Vec<u8>>> {
-        Ok(Some(vec![]))
-    }
-
-    fn deserialize(
-        &self,
-        _metadata: &[u8],
-        _session: &VortexSession,
-    ) -> VortexResult<Self::Options> {
-        Ok(EmptyOptions)
-    }
-
-    fn arity(&self, _options: &Self::Options) -> Arity {
-        Arity::Exact(2)
-    }
-
-    fn child_name(&self, _instance: &Self::Options, child_idx: usize) -> ChildName {
-        match child_idx {
-            0 => ChildName::from("list"),
-            1 => ChildName::from("needle"),
-            _ => unreachable!(
-                "Invalid child index {} for ListContains expression",
-                child_idx
-            ),
-        }
-    }
-    fn fmt_sql(
-        &self,
-        _options: &Self::Options,
-        expr: &Expression,
-        f: &mut Formatter<'_>,
-    ) -> std::fmt::Result {
-        write!(f, "contains(")?;
-        expr.child(0).fmt_sql(f)?;
-        write!(f, ", ")?;
-        expr.child(1).fmt_sql(f)?;
-        write!(f, ")")
-    }
-
-    fn return_dtype(&self, _options: &Self::Options, arg_dtypes: &[DType]) -> VortexResult<DType> {
-        let list_dtype = &arg_dtypes[0];
-        let needle_dtype = &arg_dtypes[0];
-
-        let nullability = match list_dtype {
-            DType::List(_, list_nullability) => list_nullability,
-            _ => {
-                vortex_bail!(
-                    "First argument to ListContains must be a List, got {:?}",
-                    list_dtype
-                );
-            }
-        }
-        .bitor(needle_dtype.nullability());
-
-        Ok(DType::Bool(nullability))
-    }
-
-    fn execute(&self, _options: &Self::Options, args: ExecutionArgs) -> VortexResult<ArrayRef> {
-        let [list_array, value_array]: [ArrayRef; _] = args
-            .inputs
-            .try_into()
-            .map_err(|_| vortex_err!("Wrong number of arguments for ListContains expression"))?;
-
-        if let Some(list_scalar) = list_array.as_constant()
-            && let Some(value_scalar) = value_array.as_constant()
-        {
-            let result = compute_contains_scalar(&list_scalar, &value_scalar)?;
-            return Ok(ConstantArray::new(result, args.row_count).into_array());
-        }
-
-        compute_list_contains(list_array.as_ref(), value_array.as_ref())?.execute(args.ctx)
-    }
-
-    fn stat_falsification(
-        &self,
-        _options: &Self::Options,
-        expr: &Expression,
-        catalog: &dyn StatsCatalog,
-    ) -> Option<Expression> {
-        let list = expr.child(0);
-        let needle = expr.child(1);
-
-        // falsification(contains([1,2,5], x)) =>
-        //   falsification(x != 1) and falsification(x != 2) and falsification(x != 5)
-        let min = list.stat_min(catalog)?;
-        let max = list.stat_max(catalog)?;
-        // If the list is constant when we can compare each element to the value
-        if min == max {
-            let list_ = min
-                .as_opt::<Literal>()
-                .and_then(|l| l.as_list_opt())
-                .and_then(|l| l.elements())?;
-            if list_.is_empty() {
-                // contains([], x) is always false.
-                return Some(lit(true));
-            }
-            let value_max = needle.stat_max(catalog)?;
-            let value_min = needle.stat_min(catalog)?;
-
-            return and_collect(list_.iter().map(move |v| {
-                or(
-                    lt(value_max.clone(), lit(v.clone())),
-                    gt(value_min.clone(), lit(v.clone())),
-                )
-            }));
-        }
-
-        None
-    }
-
-    // Nullability matters for contains([], x) where x is false.
-    fn is_null_sensitive(&self, _instance: &Self::Options) -> bool {
-        true
-    }
-
-    fn is_fallible(&self, _options: &Self::Options) -> bool {
-        false
-    }
-}
-
-fn compute_contains_scalar(list: &Scalar, needle: &Scalar) -> VortexResult<Scalar> {
-    let nullability = list.dtype().nullability() | needle.dtype().nullability();
-
-    // Handle null list or null needle
-    if list.is_null() || needle.is_null() {
-        return Ok(Scalar::null(DType::Bool(nullability)));
-    }
-
-    let list_scalar = list.as_list();
-    let elements = list_scalar
-        .elements()
-        .ok_or_else(|| vortex_err!("Expected non-null list"))?;
-
-    let contains = elements.iter().any(|elem| elem == needle);
-    Ok(Scalar::bool(contains, nullability))
-}
-
-/// Creates an expression that checks if a value is contained in a list.
-///
-/// Returns a boolean array indicating whether the value appears in each list.
-///
-/// ```rust
-/// # use vortex_array::expr::{list_contains, lit, root};
-/// let expr = list_contains(root(), lit(42));
-/// ```
-pub fn list_contains(list: Expression, value: Expression) -> Expression {
-    ListContains.new_expr(EmptyOptions, [list, value])
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use vortex_buffer::BitBuffer;
-    use vortex_dtype::DType;
-    use vortex_dtype::Field;
-    use vortex_dtype::FieldPath;
-    use vortex_dtype::FieldPathSet;
-    use vortex_dtype::Nullability;
-    use vortex_dtype::PType::I32;
-    use vortex_dtype::StructFields;
-    use vortex_utils::aliases::hash_map::HashMap;
-    use vortex_utils::aliases::hash_set::HashSet;
-
-    use super::list_contains;
-    use crate::Array;
-    use crate::ArrayRef;
-    use crate::IntoArray;
-    use crate::arrays::BoolArray;
-    use crate::arrays::ListArray;
-    use crate::arrays::PrimitiveArray;
-    use crate::expr::exprs::binary::and;
-    use crate::expr::exprs::binary::gt;
-    use crate::expr::exprs::binary::lt;
-    use crate::expr::exprs::binary::or;
-    use crate::expr::exprs::get_item::col;
-    use crate::expr::exprs::get_item::get_item;
-    use crate::expr::exprs::literal::lit;
-    use crate::expr::exprs::root::root;
-    use crate::expr::pruning::checked_pruning_expr;
-    use crate::expr::stats::Stat;
-    use crate::scalar::Scalar;
-    use crate::validity::Validity;
-
-    fn test_array() -> ArrayRef {
-        ListArray::try_new(
-            PrimitiveArray::from_iter(vec![1, 1, 2, 2, 2, 2, 2, 3, 3, 3]).into_array(),
-            PrimitiveArray::from_iter(vec![0, 5, 10]).into_array(),
-            Validity::AllValid,
-        )
-        .unwrap()
-        .into_array()
-    }
-
-    #[test]
-    pub fn test_one() {
-        let arr = test_array();
-
-        let expr = list_contains(root(), lit(1));
-        let item = arr.apply(&expr).unwrap();
-
-        assert_eq!(
-            item.scalar_at(0).unwrap(),
-            Scalar::bool(true, Nullability::Nullable)
-        );
-        assert_eq!(
-            item.scalar_at(1).unwrap(),
-            Scalar::bool(false, Nullability::Nullable)
-        );
-    }
-
-    #[test]
-    pub fn test_all() {
-        let arr = test_array();
-
-        let expr = list_contains(root(), lit(2));
-        let item = arr.apply(&expr).unwrap();
-
-        assert_eq!(
-            item.scalar_at(0).unwrap(),
-            Scalar::bool(true, Nullability::Nullable)
-        );
-        assert_eq!(
-            item.scalar_at(1).unwrap(),
-            Scalar::bool(true, Nullability::Nullable)
-        );
-    }
-
-    #[test]
-    pub fn test_none() {
-        let arr = test_array();
-
-        let expr = list_contains(root(), lit(4));
-        let item = arr.apply(&expr).unwrap();
-
-        assert_eq!(
-            item.scalar_at(0).unwrap(),
-            Scalar::bool(false, Nullability::Nullable)
-        );
-        assert_eq!(
-            item.scalar_at(1).unwrap(),
-            Scalar::bool(false, Nullability::Nullable)
-        );
-    }
-
-    #[test]
-    pub fn test_empty() {
-        let arr = ListArray::try_new(
-            PrimitiveArray::from_iter(vec![1, 1, 2, 2, 2]).into_array(),
-            PrimitiveArray::from_iter(vec![0, 5, 5]).into_array(),
-            Validity::AllValid,
-        )
-        .unwrap()
-        .into_array();
-
-        let expr = list_contains(root(), lit(2));
-        let item = arr.apply(&expr).unwrap();
-
-        assert_eq!(
-            item.scalar_at(0).unwrap(),
-            Scalar::bool(true, Nullability::Nullable)
-        );
-        assert_eq!(
-            item.scalar_at(1).unwrap(),
-            Scalar::bool(false, Nullability::Nullable)
-        );
-    }
-
-    #[test]
-    pub fn test_nullable() {
-        let arr = ListArray::try_new(
-            PrimitiveArray::from_iter(vec![1, 1, 2, 2, 2]).into_array(),
-            PrimitiveArray::from_iter(vec![0, 5, 5]).into_array(),
-            Validity::Array(BoolArray::from(BitBuffer::from(vec![true, false])).into_array()),
-        )
-        .unwrap()
-        .into_array();
-
-        let expr = list_contains(root(), lit(2));
-        let item = arr.apply(&expr).unwrap();
-
-        assert_eq!(
-            item.scalar_at(0).unwrap(),
-            Scalar::bool(true, Nullability::Nullable)
-        );
-        assert!(!item.is_valid(1).unwrap());
-    }
-
-    #[test]
-    pub fn test_return_type() {
-        let scope = DType::Struct(
-            StructFields::new(
-                ["array"].into(),
-                vec![DType::List(
-                    Arc::new(DType::Primitive(I32, Nullability::NonNullable)),
-                    Nullability::Nullable,
-                )],
-            ),
-            Nullability::NonNullable,
-        );
-
-        let expr = list_contains(get_item("array", root()), lit(2));
-
-        // Expect nullable, although scope is non-nullable
-        assert_eq!(
-            expr.return_dtype(&scope).unwrap(),
-            DType::Bool(Nullability::Nullable)
-        );
-    }
-
-    #[test]
-    pub fn list_falsification() {
-        let expr = list_contains(
-            lit(Scalar::list(
-                Arc::new(DType::Primitive(I32, Nullability::NonNullable)),
-                vec![1.into(), 2.into(), 3.into()],
-                Nullability::NonNullable,
-            )),
-            col("a"),
-        );
-
-        let (expr, st) = checked_pruning_expr(
-            &expr,
-            &FieldPathSet::from_iter([
-                FieldPath::from_iter([Field::Name("a".into()), Field::Name("max".into())]),
-                FieldPath::from_iter([Field::Name("a".into()), Field::Name("min".into())]),
-            ]),
-        )
-        .unwrap();
-
-        assert_eq!(
-            &expr,
-            &and(
-                and(
-                    or(lt(col("a_max"), lit(1i32)), gt(col("a_min"), lit(1i32)),),
-                    or(lt(col("a_max"), lit(2i32)), gt(col("a_min"), lit(2i32)),)
-                ),
-                or(lt(col("a_max"), lit(3i32)), gt(col("a_min"), lit(3i32)),)
-            )
-        );
-
-        assert_eq!(
-            st.map(),
-            &HashMap::from_iter([(
-                FieldPath::from_name("a"),
-                HashSet::from([Stat::Min, Stat::Max])
-            )])
-        );
-    }
-
-    #[test]
-    pub fn test_display() {
-        let expr = list_contains(get_item("tags", root()), lit("urgent"));
-        assert_eq!(expr.to_string(), "contains($.tags, \"urgent\")");
-
-        let expr2 = list_contains(root(), lit(42));
-        assert_eq!(expr2.to_string(), "contains($, 42i32)");
-    }
-
-    #[test]
-    pub fn test_constant_scalars() {
-        let arr = test_array();
-
-        // Both list and needle are constants - should use scalar optimization
-        let list_scalar = Scalar::list(
-            Arc::new(DType::Primitive(I32, Nullability::NonNullable)),
-            vec![1.into(), 2.into(), 3.into()],
-            Nullability::NonNullable,
-        );
-
-        // Test contains true
-        let expr = list_contains(lit(list_scalar.clone()), lit(2i32));
-        let result = arr.apply(&expr).unwrap();
-        assert_eq!(
-            result.scalar_at(0).unwrap(),
-            Scalar::bool(true, Nullability::NonNullable)
-        );
-
-        // Test contains false
-        let expr = list_contains(lit(list_scalar), lit(42i32));
-        let result = arr.apply(&expr).unwrap();
-        assert_eq!(
-            result.scalar_at(0).unwrap(),
-            Scalar::bool(false, Nullability::NonNullable)
-        );
-    }
-}
